imagemanager/requirements.txt
PyYAML
oslo.config
oslo.log
requests
shade
